+++ src/components/CoreComponents/Tooltip.vue
@@ -61,7 +61,7 @@ function onEnd() {
 	<div
 		:id="tooltipName"
 		ref="tooltipSlot"
-		class="bg-[#505050] rounded-md px-2 py-1 tooltip"
+		class="bg-[#505050] rounded-md px-2 py-1 tooltip z-10"
 		role="tooltip"
 	>
 		<slot name="tooltip"></slot>